+++ test/system_tests/environments/gymnasium/translational_inspection/inspection_v0_episode_data.pkl
