+++ test/system_tests/environments/gymnasium/translational_inspection/model.onnx
