+++ test/system_tests/environments/gymnasium/weighted_six_dof_inspection/model.onnx
