+++ test/system_tests/environments/gymnasium/weighted_six_dof_inspection/weighted_six_dof_inspection_v0_episode_data.pkl
