+++ test/system_tests/environments/gymnasium/weighted_translational_inspection/model.onnx
