+++ test/system_tests/environments/gymnasium/weighted_translational_inspection/weighted_inspection_v0_episode_data.pkl
