+++ test/system_tests/environments/pettingzoo/multiagent_docking/models/blue0_ctrl/model.onnx
