+++ test/system_tests/environments/pettingzoo/multiagent_docking/models/blue1_ctrl/model.onnx
