+++ test/system_tests/environments/pettingzoo/multiagent_docking/models/blue2_ctrl/model.onnx
