+++ test/system_tests/environments/pettingzoo/multiagent_translational_inspection/models/blue0_ctrl/model.onnx
