+++ test/system_tests/environments/pettingzoo/multiagent_translational_inspection/models/blue1_ctrl/model.onnx
