+++ test/system_tests/environments/pettingzoo/multiagent_weighted_six_dof_inspection/models/blue0_ctrl/model.onnx
