+++ test/system_tests/environments/pettingzoo/multiagent_weighted_six_dof_inspection/models/blue2_ctrl/model.onnx
