+++ test/system_tests/environments/pettingzoo/multiagent_weighted_translational_inspection/models/blue0_ctrl/model.onnx
