+++ test/system_tests/environments/pettingzoo/multiagent_weighted_translational_inspection/models/blue1_ctrl/model.onnx
