+++ test/system_tests/environments/pettingzoo/multiagent_weighted_translational_inspection/models/blue2_ctrl/model.onnx
